Delete lyrics on error
When there's a network error, or the lyrics can't be found, the network
reply and lyrics object need to be deleted from the lyrics map. This is
necessary because the network reply will be deleted at the end of
onGetLyricPageResult() anyway, so can no longer be used as a key in the
map.

Removing the key should prevent the app crashing when the provider is
changed after an unsuccessful search.